+++ src/modals/ValidatorDecentralization/index.tsx
@@ -15,7 +15,7 @@ import { Title } from 'library/Modal/Title';
 import { StatusLabel } from 'library/StatusLabel';
 import type { ValidatorDetail } from '@polkawatch/ddp-client';
 import { useOverlay } from '@polkadot-cloud/react/hooks';
-import { usePolkawatchApi } from '../../contexts/Polkawatch';
+import { usePolkawatchApi } from '../../contexts/Plugins/Polkawatch';
 import { PolkawatchButton } from '../../library/PolkawatchButton';
 import { usePlugins } from '../../contexts/Plugins';
